requirements.txt
flask
face_recognition
numpy
faiss-cpu
packaging